+++ frontend/src/pages/Contact.js
@@ -3,6 +3,8 @@ import React from "react";
 //******* Import component(s) *******/
 import FooterComponent from "../components/FooterComponent";
 import Navbar from "../components/NavBar";
+import ContactUs from "../components/ContactUs"
+import "../styles/CSS/contact.css"; // Import the CSS file for styling
 
 //******* Import style logic *******/
 //Here
@@ -15,6 +17,7 @@ export default function Contact() {
         <h1>Contact page </h1>
         <h2>Dig into you records and share with others!</h2>
       </div>
+      <ContactUs />
       <FooterComponent />
     </div>
   );